move helpers functions into util
