+++ client/modules/cm-tarnation/src/grammar/rules/pattern.ts
@@ -1,5 +1,5 @@
+import { isRegExpString } from "../../util"
 import type * as DF from "../definition"
-import { isRegExpString } from "../helpers"
 import { RegExpMatcher } from "../matchers/regexp"
 import { StringMatcher } from "../matchers/string"
 import type { Repository } from "../repository"
