+++ client/modules/cm-tarnation/src/index.ts
@@ -1,3 +1,2 @@
-export * from "./grammar/helpers"
 export * from "./language"
 export * from "./util"
